Move `eorzeaDatabase` to its own package
