+++ sam/interact/src/main/kotlin/Handler.kt
@@ -6,7 +6,7 @@ import cloud.drakon.ktdiscord.interaction.InteractionJsonSerializer
 import cloud.drakon.ktdiscord.interaction.applicationcommand.ApplicationCommandData
 import cloud.drakon.tempestbot.interact.api.openai.OpenAI
 import cloud.drakon.tempestbot.interact.commands.citations.citationHandler
-import cloud.drakon.tempestbot.interact.commands.ffxiv.eorzeaDatabase
+import cloud.drakon.tempestbot.interact.commands.ffxiv.eorzeadatabase.eorzeaDatabase
 import cloud.drakon.tempestbot.interact.commands.ffxiv.lodestone.lodestoneHandler
 import cloud.drakon.tempestbot.interact.commands.ffxiv.universalis
 import cloud.drakon.tempestbot.interact.commands.openai.chat
